+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
 ## Unreleased
+### Added
+* Customizable player and enemy classes #76
 
 ### Removed
 * Backwards compatibility code for binary game data from v0.4.0 #75
+
+### Changed
+* `rpg reset --hard` removes datafile instead of entire .rpg dir 5adfb87
 
 ## [0.5.0](https://github.com/facundoolano/rpg-cli/releases/tag/0.5.0) - 2021-06-26
 ### Added